// hotelpg/src/app/models/Huesped.ts
export interface Huesped {
    id_huesped?: number;
    nombres?: string;
    primer_apellido?: string;
    segundo_apellido?: string;
    documento_identidad?: string;
    fecha_nacimiento?: string;
    estado_civil?: string;
    ocupacion?: string;
    nacionalidad?: string;
    telefono?: string;
    correo?: string;
    estado?: string;
    last_update?: string;
    id_usuario_op?: number;
    id_hotel?: number;
    id_usuario_cliente?: number;
}